Fix CLI commands
