+++ generator/common/credentials.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from generator.common import const
+from common import const
 
 
 def convert_to_vp(creds: List[dict]) -> dict:
